Deferred content returns placeholder
This change is more prep to support "classic" (two-part) streaming. It adds a
placeholder for content that is deferred and adds structure to infer whether
the content should be deferred based on whether "props" returned from gSSP is
a promise. However, it still enforces that deferring content is always turned
off (until the feature is fully implemented).

See more of the full design here:
https://docs.google.com/document/d/1O60XZxvkwx4ch8D7pi67K3RLN8OtbBEbYS8qyS0PQ2w/preview#